Moved Intent-related Operations to new Package
